sparate view and logic Order Customer
and change model path, name
