+++ src/hooks/useOrder.ts
@@ -1,16 +1,23 @@
 import { useState, useCallback } from 'react';
+import { useSnackbar } from 'notistack';
+import { useTranslation } from 'react-i18next';
+import { Zoom } from '@mui/material';
 import request from 'src/hooks/useRequest';
 import type { Order } from 'src/models/order';
 import {
-  setOrders,
+  setOrders as sO,
   updateOrder,
   useOrderDispatch
 } from 'src/contexts/OrderContext';
 
 export const useOrder = () => {
+  const { t }: { t: any } = useTranslation();
+  const [orders, setOrders] = useState<Order[]>([]);
+  const [totalOrderCount, setTotalOrderCount] = useState(0);
   const [order, setOrder] = useState<Order>();
   const [loading, setLoading] = useState<boolean>(false);
   const orderDispatch = useOrderDispatch();
+  const { enqueueSnackbar } = useSnackbar();
 
   const getOrder = useCallback((orderId: number) => {
     setLoading(true);
@@ -30,6 +37,45 @@ export const useOrder = () => {
     }
   }, []);
 
+  const getOrders = useCallback((page, limit) => {
+    try {
+      request({
+        url: '/v1/orders/search',
+        method: 'GET',
+        reqParams: {
+          params: {
+            offset: page * limit,
+            limit,
+            sort_column: 'date_of_visit',
+            sort_by: 'desc'
+          }
+        }
+      }).then((response) => {
+        setOrders(response.data.orders);
+        setTotalOrderCount(response.data.totalCount);
+      });
+    } catch (err) {
+      console.error(err);
+    }
+  }, []);
+
+  const deleteOrder = useCallback((deleteId: number) => {
+    request({
+      url: `/v1/orders/${deleteId}`,
+      method: 'DELETE'
+    }).then(() => {
+      setOrders((prev) => prev.filter((o) => o.id !== deleteId));
+      enqueueSnackbar(t('The order has been removed'), {
+        variant: 'success',
+        anchorOrigin: {
+          vertical: 'top',
+          horizontal: 'right'
+        },
+        TransitionComponent: Zoom
+      });
+    });
+  }, []);
+
   const getTodayReservationOrders = useCallback(
     (page: number, limit: number) => {
       request({
@@ -43,9 +89,7 @@ export const useOrder = () => {
           }
         }
       }).then((response) => {
-        orderDispatch(
-          setOrders(response.data.orders, response.data.totalCount)
-        );
+        orderDispatch(sO(response.data.orders, response.data.totalCount));
       });
     },
     []
@@ -74,11 +118,16 @@ export const useOrder = () => {
   }, []);
 
   return {
+    getOrders,
     getOrder,
     putOrderStatus,
     putOrderTime,
     getTodayReservationOrders,
+    deleteOrder,
+
+    orders,
     order,
+    totalOrderCount,
     loading
   };
 };
