+++ src/models/administrator.ts
@@ -1,0 +1,8 @@
+export interface Administrator {
+  id: number;
+  name: string;
+  email: string;
+  fontRate: number;
+  avatar: string;
+  jobtitle: string;
+}
