--- src/models/administrators.ts
@@ -1,5 +1,0 @@
-export interface Administrators {
-  id: number;
-  name: string;
-  email: string;
-}
